pyproject.toml
[tool.poetry]
name = "duello"
version = "0.1.0"
description = "Knowledge battles WebApp"
authors = ["Victor Pereira"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
python-decouple = "^3.6"
PyJWT = "^2.3.0"
cryptography = "^36.0.1"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
ipython = "^8.0.1"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
